Rename NB Classifier to Detect Credit Card Fraud.ipynb to Naive Bayes Classifier to Detect Credit Card Fraud.ipynb
Naive Bayes Classier to Detect Credit Card Fraud

In this question you are asked to t Naive Bayes Classier to
detect credit card fraud using de-identified credit card transactions labeled as fraudulent
or genuine. The below description was adopted from kaggle website.
The datasets contain transactions made by credit cards in September 2013 by European
cardholders. This dataset presents transactions that occurred in two days, where
we have 492 frauds out of 284,807 transactions. The dataset is highly unbalanced, the
positive class (frauds) account for 0.172% of all transactions.
It contains only numerical input variables which are the result of a PCA transformation.
Unfortunately, due to condentiality issues, we cannot provide the original
features and more background information about the data. Features V1, V2, ... V28 are
the principal components obtained with PCA, the only features which have not been
transformed with PCA are `Time' and `Amount'. Feature `Time' contains the seconds
elapsed between each transaction and the rst transaction in the dataset. The feature
`Amount' is the transaction Amount, this feature can be used for example-dependent
cost-sensitive learning. Feature `Class' is the response variable (target) and it takes value
1 in case of fraud and 0 otherwise.
Please treat V1 to V28 as continuous features and use Gaussian distribution for
training as I demonstrated in the lecture, i.e., P(V 1 j Class) is Gaussian, to calculation
the class-conditional probabilities in training. And treat `Time' and `Amount' as discrete
features by discretizing them into 2 states for the former and 3 states for the latter, for
example, Low/Medium/High and Near/Far. You should divide the data into a training
set (80%) and a test set (20%) and calculate precision, recall and F-score.